many changes

+++ frontend/admin/src/actions/auth.actions.jsx
@@ -13,8 +13,8 @@ export const login = (user) => {
 
     if (res.status === 200) {
       const { token, user } = res.data;
-      localStorage.setItem("token", token);
-      localStorage.setItem("user", JSON.stringify(user));
+      localStorage.setItem('token', token);
+      localStorage.setItem('user', JSON.stringify(user));
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
@@ -35,32 +35,43 @@ export const login = (user) => {
 
 export const isUserLoggedIn = () => {
   return async (dispatch) => {
-    const token = localStorage.getItem("token");
+    const token = localStorage.getItem('token');
     if (token) {
-      const user = JSON.parse(localStorage.getItem("user"));
+      const user = JSON.parse(localStorage.getItem('user'));
       dispatch({
-          type: authConstants.LOGIN_SUCCESS,
-          payload: {
-            token,
-            user,
-          },
-       
+        type: authConstants.LOGIN_SUCCESS,
+        payload: {
+          token,
+          user,
+        },
       });
     } else {
-            dispatch({
-              type: authConstants.LOGIN_FAILURE,
-              payload: { error: 'Fallo al inciar sesión' },
-            });
-          
+      dispatch({
+        type: authConstants.LOGIN_FAILURE,
+        payload: { error: "Fallo al inciar sesión" },
+      });
     }
   };
 };
 
 export const signout = () => {
-  return async dispatch => {
-    localStorage.clear()
+  return async (dispatch) => {
+
     dispatch({
       type: authConstants.LOGOUT_REQUEST
     })
-  }
-}
+    const res = await axios.post(`/admin/signout`);
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      });
+    } else {
+        dispatch({
+          type: authConstants.LOGOUT_FAILURE,
+          payload: {error: res.data.error}
+        })
+    }
+  };
+};
